add|update
add fakeController in bot
add martiAngelController in bot
update httpController from [] to 0
update tooController add loseBot
update historyBot
updaete setting
change value from decimal to BigInt

+++ database/migrations/2021_01_19_095124_create_queues_table.php
@@ -17,7 +17,7 @@ class CreateQueuesTable extends Migration
       $table->timestamps();
       $table->bigInteger('user_id')->index();
       $table->enum('type', ['it', 'buy_wall', 'sponsor']);
-      $table->decimal('value', 12, 8);
+      $table->bigInteger('value');
       $table->boolean('send');
     });
   }
